Hard drive backup to ZIP for Task Scheduler
A revised code that allows for the integration/usage with Windows Task Scheduler.  This allows for the opening and usage of this script automatically on a monthly/bi-monthly basis without having the user to physically open and run the file.

Also, this script excludes the syncing of the 'archive' folder within my hard drive, which contains a backup of my personal computer--this file is too large to backup (or is at least too large for me to want to back it up).